+++ projectroles/plugins.py
@@ -177,6 +177,16 @@ class ProjectAppPluginPoint(PluginPoint):
         # TODO: Implement this in your app plugin
         return None
 
+    def get_statistics(self):
+        """
+        Return app statistics as a dict. Should take the form of
+        {id: {label, value, url (optional), description (optional)}}.
+
+        :return: Dict
+        """
+        # TODO: Implement this in your app plugin
+        return {}
+
 
 class BackendPluginPoint(PluginPoint):
     """Projectroles plugin point for registering backend apps"""
@@ -197,6 +207,16 @@ class BackendPluginPoint(PluginPoint):
         """Return API entry point object."""
         # TODO: Implement this in your backend plugin
         raise NotImplementedError
+
+    def get_statistics(self):
+        """
+        Return backend statistics as a dict. Should take the form of
+        {id: {label, value, url (optional), description (optional)}}.
+
+        :return: Dict
+        """
+        # TODO: Implement this in your backend plugin
+        return {}
 
 
 class SiteAppPluginPoint(PluginPoint):