+++ siteinfo/tests/test_permissions.py
@@ -1,0 +1,29 @@
+"""Tests for permissions in the siteinfo app"""
+
+from django.urls import reverse
+
+# Projectroles dependency
+from projectroles.tests.test_permissions import TestPermissionBase
+
+
+class TestSiteInfoPermissions(TestPermissionBase):
+    """Tests for siteinfo view permissions"""
+
+    def setUp(self):
+        # Create users
+        self.superuser = self.make_user('superuser')
+        self.superuser.is_superuser = True
+        self.superuser.is_staff = True
+        self.superuser.save()
+
+        self.regular_user = self.make_user('regular_user')
+
+        # No user
+        self.anonymous = None
+
+    def test_site_info(self):
+        url = reverse('siteinfo:info')
+        good_users = [self.superuser]
+        bad_users = [self.anonymous, self.regular_user]
+        self.assert_render200_ok(url, good_users)
+        self.assert_redirect(url, bad_users)